+++ loja/resources/views/vendedores/index.blade.php
@@ -8,9 +8,9 @@
             <h2>Vendedores</h2>
         </div>
         <div class="pull-right">
-
-            <a class="btn btn-success" href="{{ route('vendedores.create') }}"> + Novo Vendedor</a>
-
+            @can('vendedores-create')
+                <a class="btn btn-success" href="{{ route('vendedores.create') }}"> + Novo Vendedor</a>
+            @endcan
         </div>
     </div>
 </div>
@@ -43,19 +43,19 @@
     <td>
        <a class="btn btn-info" href="{{ route('vendedores.show',$vendedor->id) }}">Mostrar</a>
 
-       {{-- @can('vendedor-edit') --}}
+       @can('vendedor-edit')
 
         <a class="btn btn-primary" href="{{ route('vendedores.edit',$vendedor->id) }}">Editar</a>
 
-       {{-- @endcan --}}
+       @endcan
 
+        @can('vendedor-delete')
+            {!! Form::open(['method' => 'DELETE','route' => ['vendedores.destroy', $vendedor->id],'style'=>'display:inline']) !!}
 
-        {!! Form::open(['method' => 'DELETE','route' => ['vendedores.destroy', $vendedor->id],'style'=>'display:inline']) !!}
+                {!! Form::submit('Apagar', ['class' => 'btn btn-danger']) !!}
 
-            {!! Form::submit('Apagar', ['class' => 'btn btn-danger']) !!}
-
-        {!! Form::close() !!}
-
+            {!! Form::close() !!}
+        @endcan
 
     </td>
   </tr>
